requirements.txt
requests==2.31.0
feedparser==6.0.11
pyyaml==6.0.1
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
wikipedia==1.4.0
sseclient-py==1.8.0
aiohttp==3.9.1
python-dotenv==1.0.0
schedule==1.2.0
slack-sdk==3.26.0
flask==3.0.0
